+++ plugins/native_dio_adapter/lib/src/conversion_layer_adapter.dart
@@ -24,7 +24,7 @@ class ConversionLayerAdapter implements HttpClientAdapter {
   ) async {
     final request = await _fromOptionsAndStream(options, requestStream);
     final response = await client.send(request);
-    return response.toDioResponseBody();
+    return response.toDioResponseBody(options);
   }
 
   @override
@@ -41,7 +41,12 @@ class ConversionLayerAdapter implements HttpClientAdapter {
 
     request.headers.addAll(
       Map.fromEntries(
-        options.headers.entries.map((e) => MapEntry(e.key, e.value.toString())),
+        options.headers.entries.map(
+          (e) => MapEntry(
+            options.preserveHeaderCase ? e.key : e.key.toLowerCase(),
+            e.value.toString(),
+          ),
+        ),
       ),
     );
 
@@ -67,8 +72,13 @@ class ConversionLayerAdapter implements HttpClientAdapter {
 }
 
 extension on StreamedResponse {
-  ResponseBody toDioResponseBody() {
-    final dioHeaders = headers.entries.map((e) => MapEntry(e.key, [e.value]));
+  ResponseBody toDioResponseBody(RequestOptions options) {
+    final dioHeaders = headers.entries.map(
+      (e) => MapEntry(
+        options.preserveHeaderCase ? e.key : e.key.toLowerCase(),
+        [e.value],
+      ),
+    );
     return ResponseBody(
       stream.cast<Uint8List>(),
       statusCode,